Alerting: Remove redundant applies from Alertmanager interface
This PR replaces the redundant Alertmanager interface methods:
- SaveAndApplyDefaultConfig
- SaveAndApplyConfig
- ApplyConfig

With:
- SaveConfig
- ApplyConfig

In doing so, all config applies are consolidated to the same code logic flow
which makes the alertmanager struct much simpler to follow.

While doing this, a few bugs came to light around whether we store the
PostableUserConfig or the PostableApiAlertingConfig in the database and which we
 use as the hash for the in-memory alertmanager and the hash for the database
 configuration. Previously, this could vary depending on whether it was saved
 from default, loaded via SyncAlertmanagersForOrgs, or saved from the API.

This is all much clearer now and we always:

- Store the PostableUserConfig in the database alongside a hash of the
PostableUserConfig.
- Provide the PostableApiAlertingConfig to the in-memory alertmanager alongside
a hash of the PostableApiAlertingConfig.

The main technical hurdle of this change is how to replicate the
GrafanaAlertmanager config write-lock across both ApplyConfig and SaveConfig
individually when SaveConfig calls ApplyConfig in a callback. Without re-entrant
locking, this is not simple. However, I believe write locking the alertmanager
config across the entire Save->Apply transaction is unnecessary. The existing
logic that has this locking already does not succeed in preventing race
conditions between API calls and SyncAlertmanagersForOrgs. Furthermore, any race
 conditions that result in discrepencies between database and in-memory state
 will quickly be rectified on the next run of SyncAlertmanagersForOrgs.